requirements.txt
requests==2.31.0
loguru==0.7.2
python-dotenv==1.0.0
jira==3.5.2
rich==13.7.0 